e2e_testing/requirements.txt
pytest-asyncio
pytest-playwright
pillow